agents/requirements.txt
google-adk
google-genai
openai
pinecone-client
python-dotenv

langchain
langchain-community
langchain-openai
langchain-pinecone
langchain-text-splitters

tiktoken
unstructured
pypdf
pandas
numpy

uvicorn
fastapi
